Added "Identify Staging Areas" setup accordion item
