+++ app/client/src/config/options.ts
@@ -1,0 +1,4 @@
+export const hazardousOptions: { label: string; value: string }[] = [
+  { label: 'Hazardous', value: 'hazardous' },
+  { label: 'Non-Hazardous', value: 'non-hazardous' },
+];
