+++ app/client/src/contexts/LookupFiles.tsx
@@ -176,6 +176,7 @@ type Content = {
     gpServerInputMaxRecordCount: number;
     proxyUrl: string;
     structures: string;
+    suitability: string;
     totsGPServer: string;
     totsTestGPServer: string;
     useProxyForGPServer: boolean;
