+++ app/client/src/contexts/Sketch.tsx
@@ -32,11 +32,7 @@ import {
 } from 'config/sampleAttributes';
 // config
 import { isDecon } from 'config/navigation';
-
-export const hazardousOptions: { label: string; value: string }[] = [
-  { label: 'Hazardous', value: 'hazardous' },
-  { label: 'Non-Hazardous', value: 'non-hazardous' },
-];
+import { hazardousOptions } from 'config/options';
 
 type HomeWidgetType = {
   '2d': __esri.Home;
@@ -209,6 +205,8 @@ type SketchType = {
   setResultsOpen: Dispatch<SetStateAction<boolean>>;
   efficacyResults: any;
   setEfficacyResults: Dispatch<SetStateAction<any>>;
+  suitabilityLayerVisible: boolean;
+  setSuitabilityLayerVisible: Dispatch<SetStateAction<boolean>>;
 };
 
 export const SketchContext = createContext<SketchType>({
@@ -302,6 +300,8 @@ export const SketchContext = createContext<SketchType>({
   setResultsOpen: () => {},
   efficacyResults: null,
   setEfficacyResults: () => {},
+  suitabilityLayerVisible: true,
+  setSuitabilityLayerVisible: () => {},
 });
 
 type Props = { children: ReactNode };
@@ -414,6 +414,7 @@ export function SketchProvider({ children }: Props) {
   const [viewUnderground3d, setViewUnderground3d] = useState(false);
   const [resultsOpen, setResultsOpen] = useState(false);
   const [efficacyResults, setEfficacyResults] = useState(null);
+  const [suitabilityLayerVisible, setSuitabilityLayerVisible] = useState(true);
 
   // Update totsLayers variable on the window object. This is a workaround
   // to an issue where the layers state variable is not available within esri
@@ -726,6 +727,8 @@ export function SketchProvider({ children }: Props) {
         setResultsOpen,
         efficacyResults,
         setEfficacyResults,
+        suitabilityLayerVisible,
+        setSuitabilityLayerVisible,
       }}
     >
       {children}
