+++ app/server/app/public/data/config/services.json
@@ -4,6 +4,7 @@
   "gpServerInputMaxRecordCount": 2000,
   "proxyUrl": "https://app7.erg.com/PHP/proxy.php?",
   "structures": "https://services2.arcgis.com/FiaPA4ga0iQKduv3/arcgis/rest/services/USA_Structures_View/FeatureServer/0",
+  "suitability": "https://services1.arcgis.com/BN5Mj6q47j71sqyJ/arcgis/rest/services/SitingStagingNC250m/FeatureServer/1",
   "useProxyForGPServer": false,
   "radarDatasets": {
     "sampleMetadata": "https://radar-staging.app.cloud.gov/api/v1/datasets/b334f501-10be-4bc3-8fbf-7eacd163eff7/data/1"
